==> 02_DGE.Rmd <==
---
title: "DGE"
author: "tobiasst"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output:
  html_document:
    df_print: paged
    code_folding: hide
---

```{r init, message=F, warning=F}
library(SummarizedExperiment)
library(DESeq2)
library(parallel)
library(pheatmap)
```

```{r stats,  warning=F}
se <- readRDS("se.rds")
dds <- DESeqDataSet(se, design = ~ diet)
dds <- DESeq(dds)
resultsNames(dds)
res <- results(dds, contrast=c("diet","HF","LF"))
summary(res)
saveRDS(res, "res_HF_v_LF.rds")

## fits suppl. table 
rf <- data.frame(res[order(res$padj),])
rf$symbol <- rowData(dds)$symbol

## differentially expressed genes
degf <- rf[!is.na(rf$padj) & rf$padj<0.1,]
res <- res[order(res$stat, decreasing = T),]
```

## upregulated genes HF versus LF

```{r heatmap, fig.height=15, fig.width=4}

upregulated <- rownames(res)[!is.na(res$padj) & res$padj < 0.1 & res$stat > 0]

mat <- t(scale(t(log2(assays(se)$tpms[upregulated,]+0.01))))
rownames(mat) <- rowData(se[upregulated,])$symbol

ann <- data.frame(diet=se$diet,row.names = colnames(mat))
p <- pheatmap(mat, annotation_col = ann, scale = "row", cluster_rows=F)
```


```{r}
sessionInfo()
```
