==> 06_GSEA.Rmd <==
---
title: "GSEA"
author: "tobiasst"
date: '`r format(Sys.time(), "%d %B, %Y")`'
output:
  html_document:
    df_print: paged
    toc: true 
    toc_depth: 3
    number_sections: true  ## if you want number sections at each table header
    code_folding: hide
    css: share/custom.css

---

gene sets downloaded from 
http://bioinf.wehi.edu.au/software/MSigDB/index.html


```{r init, message=F, warning=F}
library(SummarizedExperiment)
library(fgsea)
library(org.Mm.eg.db)
library(pheatmap)
```

ranking on the statistic (alternative log2FC)
http://bioinf.wehi.edu.au/software/MSigDB/index.html

rank vector does not have to be ordered, watch out multiple genes with same entrez id.

```{r}
se <- readRDS("se.rds")
results <- readRDS("res_HF_v_LF.rds")

ranks <- results$stat
names(ranks) <- rowData(se)$entrez
ranks <- na.omit(ranks)
ranks <- ranks[!duplicated(names(ranks))]
```

## Hallmark gene sets

```{r, message=F, fig.width=10, fig.height=10}
load("share/mouse_H_v5.rdata")
pathways <- Mm.H
fgseaRes <- fgsea(pathways, ranks, minSize=15, maxSize = 500, nperm=1000)
topPathwaysUp <- fgseaRes[ES > 0][head(order(pval), n=10), pathway]
topPathwaysDown <- fgseaRes[ES < 0][head(order(pval), n=10), pathway]
topPathways <- c(topPathwaysUp, rev(topPathwaysDown))
plotGseaTable(pathways[topPathways], ranks, fgseaRes, 
              gseaParam = 0.5)

```


## GO gene sets

```{r, message=F, fig.width=10, fig.height=10}

load("share/mouse_c5_v5p2.rdata")
pathways <- Mm.c5
fgseaRes <- fgsea(pathways, ranks, minSize=15, maxSize = 500, nperm=1000)
topPathwaysUp <- fgseaRes[ES > 0][head(order(pval), n=10), pathway]
topPathwaysDown <- fgseaRes[ES < 0][head(order(pval), n=10), pathway]
topPathways <- c(topPathwaysUp, rev(topPathwaysDown))
plotGseaTable(pathways[topPathways], ranks, fgseaRes, 
              gseaParam = 0.5)

```

## curated gene sets

```{r, message=F, fig.width=10, fig.height=10}

load("share/mouse_c2_v5.rdata")
pathways <- Mm.c2
fgseaRes <- fgsea(pathways, ranks, minSize=15, maxSize = 500, nperm=1000)
topPathwaysUp <- fgseaRes[ES > 0][head(order(pval), n=10), pathway]
topPathwaysDown <- fgseaRes[ES < 0][head(order(pval), n=10), pathway]
topPathways <- c(topPathwaysUp, rev(topPathwaysDown))
plotGseaTable(pathways[topPathways], ranks, fgseaRes, 
              gseaParam = 0.5)

```



```{r}
sessionInfo()
```
